feat(brand-guide): export browser-free buildHTML via ./brand-guide subpath
Split the pure HTML builder out of pdf.ts into brand-guide.ts so consumers
can render a brand-guide document without pulling in the Playwright browser
loader. pdf.ts keeps generatePDF and re-exports buildHTML for the CLI.

Enables dembrandt-next to import { buildHTML } from 'dembrandt/brand-guide'
in a Vercel serverless/edge route (design.md -> tokens -> HTML -> print)
with no browser dependency in the bundle.
